Solve error in test task use case file
