+++ src/test/java/test_entities/TestCategory.java
@@ -1,4 +1,4 @@
-package testentities;
+package test_entities;
 
 import entities.Category;
 import org.junit.Test;
