+++ src/test/java/test_entities/TimerTest.java
@@ -1,4 +1,4 @@
-package testentities;
+package test_entities;
 
 public class TimerTest {
 }
